MaybeInherited: Implement `local()` fn
This roughly matches the `ok()` fn of the built-in `Result` type (or `some()` of `Option`), and allows us to access the value inside the `Local` variant in a more ergonomic way. This is especially useful in combination with the `as_ref()` fn since `foo.as_ref().local()` returns `Option<&T>` which only borrows instead of moving the data.