Fill explicit `lib` declaration with default values
Previously, it was only possible to either use the auto-discovered lib target or explicitly declare one. In reality, cargo merges the definitions though. It is for example possible to rely on the auto-discovered `src/lib.rs`, but rename the library via `lib.name`.

This commit changes the `lib = Some(_)` code path to assign proper defaults to the `name`, `path`, `edition`, and `crate_type` fields unless they are explicitly specified, which matches the behavior from cargo.