Reimplement target auto-discovery
This fixes the merging behavior of the auto-discovery code. Previously an explicit `[[bin]]` would turn off the auto-discovery unconditionally, but:

- in the 2018 edition the behavior was changed so that only `autobins = false` would disable the auto-discovery
- the auto-discovery information should still be used to fill in the `path` values

This commit changes the behavior to act accordingly. cargo shows errors in a couple of more situations (see tests), but the new implementation explicitly does not treat these as hard errors to be able to ignore broken target declarations in existing crates and turn them into warnings instead.